Created Adapter files and moved content to adapters
